+++ lib/src/exception/is_not_list_exception.dart
@@ -10,12 +10,11 @@
 /// Exception thrown when a LeoML document is not well-formed and is not a list.
 class IsNotListException implements Exception {
   @override
-  String toString() =>
-      'Your LeoML document is not well-formed.\n'
-          'It must be a list. Each LeoML document must have rectangular brackets at the root.\n\n'
-          'For example:\n\n'
-          '['
-          ' {"name":"value"},'
-          ' {"name":"value"}'
-          ']';
+  String toString() => 'Your LeoML document is not well-formed.\n'
+      'It must be a list. Each LeoML document must have rectangular brackets at the root.\n\n'
+      'For example:\n\n'
+      '['
+      ' {"name":"value"},'
+      ' {"name":"value"}'
+      ']';
 }