+++ lib/src/exception/is_not_object_exception.dart
@@ -17,4 +17,3 @@ class IsNotObjectException implements Exception {
       ' {"type":"blog"},'
       ']';
 }
-