+++ lib/src/templates/widgets/section.dart
@@ -38,4 +38,3 @@ class Section extends StatelessWidgetTemplate {
     }
   }
 }
-