+++ test/src/templates/test_content_template_impl.dart
@@ -10,7 +10,10 @@ import 'package:flutter/widgets.dart';
 import 'package:leoml_parser/src/templates/content_template.dart';
 
 class TestContentTemplateImpl extends ContentTemplate {
-  TestContentTemplateImpl({required super.type, required super.defaultWidgetFactory});
+  TestContentTemplateImpl({
+    required super.type,
+    required super.defaultWidgetFactory,
+  });
 
   @override
   bool assertLeoMLStructure(List parsedLeoMLDocument) {
@@ -24,8 +27,7 @@ class TestContentTemplateImpl extends ContentTemplate {
 
   @override
   bool hasCustomWidget({required String key}) {
-
-    if(key == 'headline'){
+    if (key == 'headline') {
       return true;
     }